requirements.txt
python-telegram-bot==13.0
requests
openai==0.27.3
